Complete rethink. Use the power of the :exports-fn feature of shadow-cljs to greatly simplify this library.
Using exports-fn we can move everything that was done in the wrapper class to an object-preparation
stage, so that the JS-exported object is "ready to use" out of the box. After the dust settled,
we weren't getting much from TypeScript, so now this is a JS-only library. It generates everything
needed into one compact JS file.